0716/수정 추가 반영

+++ src/main/java/com/mycoffeemap/bean/BeanRepository.java
@@ -17,5 +17,6 @@ public interface BeanRepository extends JpaRepository<Bean, Long> {
     List<Bean> findByFlavorNotesContainingIgnoreCase(String flavor);
     
     // 내가 등록한 원두 목록 조회
-    List<Bean> findByUserId(Integer userId);
+    List<Bean> findByUserId(Long userId);
+
 }
